Change data type from double precision to single precision(float) because of limitation of liquid-dsp libarary.

+++ mainwindow.cpp
@@ -352,7 +352,7 @@ void MainWindow::createToolbar()
 	scaleSpinBox->setDecimals(3);
 	scaleSpinBox->setMinimum(0.000001);
 	toolBar->addWidget(scaleSpinBox);
-	connect(scaleSpinBox, &QDoubleSpinBox::valueChanged, this, [=](double val){
+	connect(scaleSpinBox, &QDoubleSpinBox::valueChanged, this, [=](float val){
 			auto docker = dynamic_cast<ProfileDocker*>(mainTab->tabWidget->currentWidget());
 			if(!docker)
 				return;
